test: improve core coverage to 100%
